+++ transpiler/main/tables/localvarsymtable.h
@@ -58,4 +58,5 @@ void freeLocalVarSymTable(struct LVST* lvst);
 //obtain a reference to the line for that identifier
 struct LVSTLine* lvst_get(struct LVST* lvst, char* name);
 
+void lvst_print(struct LVST* lvst);
 #endif