+++ transpiler/main/typeinference.h
@@ -5,5 +5,5 @@
 #include "../../ast/ast.h"
 
 struct Type* inferTypeExpr(struct ST* st, struct Expr* expr);
-
+char* typeToStr(struct Type* t);
 #endif