+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Added Local Variable Type Inference
 - Added 'sd -version' Option
 - Added Valgrind Tests to detect Memory Leaks
+- Added '-clean' Flag to dragon-lexer
+- Added '-clean' Flag to sd (in transpiler/)
 
 ## [0.0.3] - 2020-08-27
 - Added 'break' Statement